+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Use exact uid cardinality for uid range partitioning. Combined with predicate partitioning, large
   predicates get split into more partitions than small predicates.
+- Improve performance of `PredicatePartitioner` for a single predicate per partition (`dgraph.partitioner.predicate.predicatesPerPartition=1`).
+  This becomes the new default for this partitioner.
 
 ### Fixed
 - Dgraph groups with no predicates caused a `NullPointerException`.